updated lexical analyzer, fixes, new program to create lexical analyzer

+++ include/cerberus/algorithms.hpp
@@ -23,7 +23,10 @@ namespace cerb {
     {
         if (std::is_constant_evaluated()) {
             if constexpr (std::is_trivial_v<T>) {
-                return memcpy(dest, begin, static_cast<long>(end - begin));
+                return memcpy(
+                    dest,
+                    begin,
+                    static_cast<size_t>(static_cast<ptrdiff_t>(end - begin)));
             }
         }
